requirements.txt
asgiref==3.7.1
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==3.2.19
djangorestframework==3.14.0
idna==3.4
isodate==0.6.1
lxml==4.9.2
platformdirs==3.5.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.4.4
suds-py3==1.4.5.0
typing_extensions==4.6.1
urllib3==2.0.2
zeep==4.2.1
